Ensured all patches do nothing if there is no Spotify Client or track info available (until they come available).
Stop throwing errors when no track is found (was causing crashes)

Made the default clientSecret and clientId be strings to make people's lives easier when following the instructions.

Removed the Retry Helper class/methods.

GetTrackInfo and RefreshSession are now called from the Jukebox UpdateStudio method, and the old SetInterval method has been removed. This fixes crashes that could occur when exiting the game.

Added a mod option to double press the stop button to return to the start of the track (the first press just pauses the song). By default it is just one press to match the Jukebox's functionality.

Exiting the game no longer resets your track to the beginning.

The Jukebox now properly updates the track name when not playing.

Fixed hitting pev or next buttons playing the song when the jukebox was stopped.

When you are standing in one place/the Jukebox volume doesn't need to change, volume update requests to Spotify are paused. This greatly reduces the number of volume requests sent to the Spotify device.