Big refactor of JukeboxPatcher to make the logic much more simple and easy to follow.
Fixed music not pausing/resuming properly when pauseOnLeave is disabled in the mod settings.

Fixed volume not getting set back to 100 when quitting SubnauticaBZ, as well as the music not pausing if you weren't listening before you started the game.

Fix for jukebox sometimes pausing when it shouldn't after going in and out of the in-game menu.

Subnautica BZ is now closed before the build runs if it is still open.